Update README.md
As a driven Python programmer, I had the privilege of undertaking an internship at Codsoft, a renowned company. During this invaluable experience, I was tasked with completing a series of challenging projects that tested my skills and pushed me to explore the depths of Python's capabilities. This repository serves as a testament to my dedication and the knowledge I acquired throughout the internship. Within these files, you'll find meticulously crafted solutions to intricate problems, each one a stepping stone towards mastering the art of Python programming. From developing a robust Rock-Paper-Scissors game and a versatile calculator to building an efficient contact manager, a user-friendly to-do list application, and a secure password generator, this repository chronicles my journey of growth and accomplishment. Dive into the code, witness the elegance of my solutions, and experience the passion that fueled every line of code.